requirements.txt
celery==5.2.7
Django==4.1.7
django-filter==23.1
django-extensions==3.2.1
django-phonenumber-field==7.0.2
django-debug-toolbar
django-admin-rangefilter
django-import-export
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
ipython

requests==2.28.2

django-crispy-forms
crispy_bootstrap5

#flake8
flake8
flake8-builtins
flake8-print

#pytest
pytest
pytest-django==4.5.2
pytest-cov==4.0.0
pytest-mock==3.10.0
